Fixed card registration and register view, fixed balance to decimal

+++ Virtual Wallet/Models/Entities/Transaction.cs
@@ -5,14 +5,14 @@
 		public Guid Id { get; set; } // специално ИД на трансакцията
 		public DateTime Timestamp { get; private set; }
 		//public string Currency { get; private set; } //Закоментирано засега, докато измислим логиката за конвертиране
-		public double Amount { get; private set; }
+		public decimal Amount { get; private set; }
 		public TransactionType Type { get; private set; } // "Add" or "Withdraw" or "Convert"
 
         // Foreign key to Wallet
         public int WalletId { get; set; }
         public Wallet Wallet { get; set; }
 
-        public Transaction(DateTime timestamp,  double amount, TransactionType type, int walletId)
+        public Transaction(DateTime timestamp, decimal amount, TransactionType type, int walletId)
         {
             Id = Guid.NewGuid();
             Timestamp = timestamp;
